+++ backend/utils/multer.js
@@ -1,5 +1,4 @@
 import multer from 'multer';
-import Datauri from 'datauri'
 import DatauriParser from 'datauri/parser.js'
 import path from 'path'
 
@@ -7,7 +6,6 @@ const storage = multer.memoryStorage()
 
 const multerUploads = multer({storage}).single('image')
 
-// const dUri = new Datauri()
 const parser = new DatauriParser()
 
 const dataUri = (req) => {parser.format(path.extname(req.file.originalname).toString(), req.file.buffer).content}
